Java inheritance I
Java inheritance I is all about understanding the idea of inheritance in oops. 
Inheritance: It is a concept in oops in which there is a super class i.e parent class whose data is inherited by another subclass i.e child class but the data members of the child class are not accessible by the parent class.  

In the following example the child class is Birds and the parent class is Animal. The child class is able to access the walk() method of the parent class but the parent class i.e Animals cannot access the methods of the child class fly() and sing().

Similarly, it happens in any kind of inheritance. The different types of inheritance are discussed in the readme file of this repository. Must give it a read.