Inheritance II
Abstraction is the concept in which we limit the access of some part of the code by the user. The abstract class is only accessible by its child class and nothing else. Also we cannot create an object for the abstract class but we can only access its data members with the help of its child class objects.